% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Discriminant.R
\name{Graph.KNN}
\alias{Graph.KNN}
\title{Graph.KNN}
\usage{
Graph.KNN(X, Y, k)
}
\arguments{
\item{X}{A data frame or a matrix where rows are observations and columns are features.}

\item{Y}{A vector with labels for each row in \code{data}.}

\item{k}{Number of neighbors that the classifier should use. It has to be an odd number.}
}
\value{
A graph that plots the boundries and how the categories are delimited.
}
\description{
Find the Discriminant Function that separates three different categories.
}
\examples{
# create artificial dataset
x1 <- c(978783,1052488,1039495,1056795,1125545,1536011,1616461,1388507,1608121,1416574,1704919,2653310,2208399,1896304,1957401)
x2 <- c(12645844,12343453,4137266,12229065, 12554668,8856611,12137668,11545424,9253718,8863474,15145969,12793921,14593861,15161586,15785243)
Y <- c("Cats","Cats","Cats","Cats","Cats","Dogs","Dogs","Dogs","Dogs","Dogs","Lions","Lions","Lions","Lions","Lions")
X <- cbind(x1,x2)
# get the kNN predictions for the test set
Graph.KNN(X=X,Y=Y,k=5)
}

