% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/kNN.R
\name{KNN.k}
\alias{KNN.k}
\title{kNN classifier}
\usage{
KNN.k(X, Y, k, obj = "train", RealData = NULL)
}
\arguments{
\item{X}{A data frame or a matrix where rows are observations and columns are features. If \code{type} is "train" this is training dataset, and if it is "predict" it is test dataset.}

\item{Y}{A vector with labels for each row in \code{data} if \code{type} is "train", and with labels for each row in \code{memory} if \code{type} is "predict".}

\item{k}{Number of neighbors that the classifier should use. It has to be an odd number.}

\item{obj}{Whether the goal is to train the classifier or predict classes of new observations based on past ones. The value can be either "train" or "predict".}

\item{RealData}{A data frame or a matrix where rows are observations and columns are features. If \code{type} is "train" this argument is not needed, and if it is "predict" it is a training dataset.}
}
\value{
A list with following elements: predictedClasses, accuracy and errorCount.
}
\description{
Classify the input with a k nearest neighbors classifier, the distance is measure as a Euclidean Distance.
}
\examples{
# create artificial dataset
inputsTest   <- matrix(rnorm(200), ncol=2)
inputsTrain  <- matrix(rnorm(200), ncol=2)
classesTrain <- c(rep(0, 50), rep(1, 50))
# get the kNN predictions for the test set
kNN_classifier(inputsTest, classesTrain, inputsTrain, k=15, obj="predict")
}

