% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN.R
\name{kNN}
\alias{kNN}
\title{kNN classifier}
\usage{
kNN(data, trueClasses, memory = NULL, k = 1, p = 2, type = "train")
}
\arguments{
\item{data}{A data frame or a matrix where rows are observations and columns are features. If \code{type} is "train" this is training dataset, and if it is "predict" it is test dataset.}

\item{trueClasses}{A vector with labels for each row in \code{data} if \code{type} is "train", and with labels for each row in \code{memory} if \code{type} is "predict".}

\item{memory}{A data frame or a matrix where rows are observations and columns are features. If \code{type} is "train" this argument is not needed, and if it is "predict" it is a training dataset.}

\item{k}{Number of neighbors that the classifier should use. It has to be an odd number.}

\item{p}{Distance metric the classifier should use, the value can be either 1, 2 or Inf.}

\item{type}{Whether the goal is to train the classifier or predict classes of new observations based on past ones. The value can be either "train" or "predict".}
}
\value{
A list with following elements: predictedClasses, prob, accuracy and errorCount.
}
\description{
Classify the input with a k nearest neighbors classifier.
}
\examples{
# create artificial dataset
inputsTest   <- matrix(rnorm(200), ncol=2)
inputsTrain  <- matrix(rnorm(200), ncol=2)
classesTrain <- c(rep(0, 50), rep(1, 50))
# get the kNN predictions for the test set
kNN(inputsTest, classesTrain, inputsTrain, k=15, p=2, type="predict")
}

