% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/SortedDist.R
\name{SortDist}
\alias{SortDist}
\title{Sorted L_n distances}
\usage{
SortDist(TrainData, TestData = NULL, p = 2)
}
\arguments{
\item{TrainData}{A data frame with n columns of features.}

\item{TestData}{A data frame with the same number of columns of TrainData. If TestData is not provided, SortDist calculates the distance between all pairs of points in TrainData.}

\item{p}{Order of the L_p distance metric to be used.}
}
\value{
A matrix with indexes of the neighbours sorted by distance. Each row corresponds to a data point in TestData (or TrainData if TestData is not provided). Each column corresponds to the kth nearest neighbour in TestData.
}
\description{
Calculate and sort the distance between all pairs of data points.
}
\examples{
Data <- mtcars[,c("mpg", "wt", "cyl")]
nTrain <- 25
p <- 2
TrainData <- Data[1:nTrain, c("mpg", "wt")]
TestData <- Data[(nTrain+1):nrow(mtcars), c("mpg", "wt")]

#Calculate the distance between elements in the Training Set
TrainSortedDist <- SortDist(TrainData, p=2)
# or Calculate distante between each element o the TestData and the training set
TestSortedDist <- SortDist(TrainData,TestData,2)
}

