% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Discriminant.R
\name{Graph.KNN}
\alias{Graph.KNN}
\title{Graph.KNN}
\usage{
Graph.KNN(X, Y, k)
}
\arguments{
\item{X}{A data frame or a matrix where rows are observations and columns are features.}

\item{Y}{A vector with labels for each row in \code{data}.}

\item{k}{Number of neighbors that the classifier should use. It has to be an odd number.}
}
\value{
A graph that plots the boundries and how the categories are delimited.
}
\description{
Find the Discriminant Function that separates three different categories.
}
\examples{
# create artificial dataset
inputsTest   <- matrix(rnorm(200), ncol=2)
inputsTrain  <- matrix(rnorm(200), ncol=2)
classesTrain <- c(rep(0, 50), rep(1, 50))
# get the kNN predictions for the test set
kNN_classifier(inputsTest, classesTrain, inputsTrain, k=15, obj="predict")
}

